feat: US-003 - Move auth service to libs with provider injection
